finetune/pyproject.toml
[tool.poetry]
name = "finetune"
version = "0.1.0"
description = ""
authors = ["Hewitt <wangh3@bxscience.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.2.1"
torch = "^2.5.1"
transformers = "^4.48.0"
datasets = "^3.2.0"
accelerate = "^1.2.1"
tqdm = "^4.67.1"
discord = "^2.3.2"
pip = "^24.3.1"
audioop-lts = "^0.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
